+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - BREAKING: The deep storage on s3 and the s3 config for ingestion have been changed to use the operator-rs commons::s3 structs ([#228])
+- `operator-rs` `0.15.0` -> `0.21.0` ([#228])
 
 [#228]: https://github.com/stackabletech/druid-operator/pull/228