+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Readiness probe added ([#241])
+- Support s3 path style access ([#245])
 
 ### Changed
 
@@ -22,6 +23,7 @@ All notable changes to this project will be documented in this file.
 [#238]: https://github.com/stackabletech/druid-operator/pull/238
 [#241]: https://github.com/stackabletech/druid-operator/pull/241
 [#244]: https://github.com/stackabletech/druid-operator/pull/244
+[#245]: https://github.com/stackabletech/druid-operator/pull/245
 
 ## [0.5.0] - 2022-03-15